Move SystemCharacteristics to a separate module
