+++ src/xpk/core/core.py
@@ -31,7 +31,6 @@ Next Steps:
     not be a universal argument.
 """
 
-import argparse
 import datetime
 import enum
 import os
@@ -43,6 +42,12 @@ import sys
 import time
 from .. import utils as xpk_utils
 from dataclasses import dataclass
+from .system_characteristics import (
+    AcceleratorType,
+    AcceleratorTypeToAcceleratorCharacteristics,
+    SystemCharacteristics,
+    get_system_characteristics,
+)
 
 ################### Compatibility Check ###################
 # Check that the user runs the below version or greater.
@@ -582,1265 +587,11 @@ autoprovisioning_custom_resource_type = """
 """
 
 
-AcceleratorType = {'TPU': 1, 'GPU': 2, 'CPU': 3}
-
-
 @dataclass
 class AutoprovisioningConfig:
   config_filename: str
   minimum_chips: int
   maximum_chips: int
-
-
-@dataclass
-class AcceleratorCharacteristics:
-  resource_type: str
-  accelerator_label: str
-  machine_label: str
-
-
-AcceleratorTypeToAcceleratorCharacteristics = {
-    # TPU
-    AcceleratorType['TPU']: AcceleratorCharacteristics(
-        'google.com/tpu',
-        'cloud.google.com/gke-tpu-accelerator',
-        'cloud.google.com/gke-tpu-topology',
-    ),
-    # GPU
-    AcceleratorType['GPU']: AcceleratorCharacteristics(
-        'nvidia.com/gpu',
-        'cloud.google.com/gke-accelerator',
-        'cloud.google.com/gce-machine-type',
-    ),
-    # CPU
-    AcceleratorType['CPU']: AcceleratorCharacteristics(
-        'cpu', '', 'cloud.google.com/gke-nodepool'
-    ),
-}
-
-
-@dataclass
-class SystemCharacteristics:
-  topology: str
-  vms_per_slice: int
-  gke_accelerator: str
-  gce_machine_type: str
-  chips_per_vm: int
-  accelerator_type: AcceleratorType  # type: ignore
-  device_type: str
-
-
-################### Subcommand Helper Functions #############################
-""" !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-IF YOU MODIFY THE BELOW UserFacingNameToSystemCharacteristics MAP YOU SHOULD
-ALSO ADD CORRESPONDING MODIFICATIONS TO UserFacingNameToSystemCharacteristics
-IN MaxText/accelerator_to_spec_map.py !!!!! """
-# vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv
-UserFacingNameToSystemCharacteristics = {
-    # GPU system characteristics
-    # A100-40gb-$CHIPS
-    'a100-40gb-1': SystemCharacteristics(
-        'N/A',
-        1,
-        'nvidia-tesla-a100',
-        'a2-highgpu-1g',
-        1,
-        AcceleratorType['GPU'],
-        'a100-40gb-1',
-    ),
-    'a100-40gb-2': SystemCharacteristics(
-        'N/A',
-        1,
-        'nvidia-tesla-a100',
-        'a2-highgpu-2g',
-        2,
-        AcceleratorType['GPU'],
-        'a100-40gb-2',
-    ),
-    'a100-40gb-4': SystemCharacteristics(
-        'N/A',
-        1,
-        'nvidia-tesla-a100',
-        'a2-highgpu-4g',
-        4,
-        AcceleratorType['GPU'],
-        'a100-40gb-4',
-    ),
-    'a100-40gb-8': SystemCharacteristics(
-        'N/A',
-        1,
-        'nvidia-tesla-a100',
-        'a2-highgpu-8g',
-        8,
-        AcceleratorType['GPU'],
-        'a100-40gb-8',
-    ),
-    # H100-80gb-$CHIPS
-    'h100-80gb-8': SystemCharacteristics(
-        'N/A',
-        1,
-        'nvidia-h100-80gb',
-        'a3-highgpu-8g',
-        8,
-        AcceleratorType['GPU'],
-        'h100-80gb-8',
-    ),
-    # H100-mega-80gb-$CHIPS
-    'h100-mega-80gb-8': SystemCharacteristics(
-        'N/A',
-        1,
-        'nvidia-h100-mega-80gb',
-        'a3-megagpu-8g',
-        8,
-        AcceleratorType['GPU'],
-        'h100-mega-80gb-8',
-    ),
-    # TPU system characteristics
-    # v5p
-    'v5p-8': SystemCharacteristics(
-        '2x2x1',
-        1,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8',
-    ),
-    'v5p-16': SystemCharacteristics(
-        '2x2x2',
-        2,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-16',
-    ),
-    'v5p-32': SystemCharacteristics(
-        '2x2x4',
-        4,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-32',
-    ),
-    'v5p-64': SystemCharacteristics(
-        '2x4x4',
-        8,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-64',
-    ),
-    'v5p-128': SystemCharacteristics(
-        '4x4x4',
-        16,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-128',
-    ),
-    'v5p-256': SystemCharacteristics(
-        '4x4x8',
-        32,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-256',
-    ),
-    'v5p-384': SystemCharacteristics(
-        '4x4x12',
-        48,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-384',
-    ),
-    'v5p-512': SystemCharacteristics(
-        '4x8x8',
-        64,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-512',
-    ),
-    'v5p-640': SystemCharacteristics(
-        '4x4x20',
-        80,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-640',
-    ),
-    'v5p-768': SystemCharacteristics(
-        '4x8x12',
-        96,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-768',
-    ),
-    'v5p-896': SystemCharacteristics(
-        '4x4x28',
-        112,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-896',
-    ),
-    'v5p-1024': SystemCharacteristics(
-        '8x8x8',
-        128,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1024',
-    ),
-    'v5p-1152': SystemCharacteristics(
-        '4x12x12',
-        144,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1152',
-    ),
-    'v5p-1280': SystemCharacteristics(
-        '4x8x20',
-        160,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1280',
-    ),
-    'v5p-1408': SystemCharacteristics(
-        '4x4x44',
-        176,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1408',
-    ),
-    'v5p-1536': SystemCharacteristics(
-        '8x8x12',
-        192,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1536',
-    ),
-    'v5p-1664': SystemCharacteristics(
-        '4x4x52',
-        208,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1664',
-    ),
-    'v5p-1792': SystemCharacteristics(
-        '4x8x28',
-        224,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1792',
-    ),
-    'v5p-1920': SystemCharacteristics(
-        '4x12x20',
-        240,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-1920',
-    ),
-    'v5p-2048': SystemCharacteristics(
-        '8x8x16',
-        256,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2048',
-    ),
-    'v5p-2176': SystemCharacteristics(
-        '4x4x68',
-        272,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2176',
-    ),
-    'v5p-2304': SystemCharacteristics(
-        '8x12x12',
-        288,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2304',
-    ),
-    'v5p-2432': SystemCharacteristics(
-        '4x4x76',
-        304,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2432',
-    ),
-    'v5p-2560': SystemCharacteristics(
-        '8x8x20',
-        320,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2560',
-    ),
-    'v5p-2688': SystemCharacteristics(
-        '4x12x28',
-        336,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2688',
-    ),
-    'v5p-2816': SystemCharacteristics(
-        '4x8x44',
-        352,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2816',
-    ),
-    'v5p-2944': SystemCharacteristics(
-        '4x4x92',
-        368,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-2944',
-    ),
-    'v5p-3072': SystemCharacteristics(
-        '8x12x16',
-        384,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3072',
-    ),
-    'v5p-3200': SystemCharacteristics(
-        '4x20x20',
-        400,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3200',
-    ),
-    'v5p-3328': SystemCharacteristics(
-        '4x8x52',
-        416,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3328',
-    ),
-    'v5p-3456': SystemCharacteristics(
-        '12x12x12',
-        432,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3456',
-    ),
-    'v5p-3584': SystemCharacteristics(
-        '8x8x28',
-        448,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3584',
-    ),
-    'v5p-3712': SystemCharacteristics(
-        '4x4x116',
-        464,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3712',
-    ),
-    'v5p-3840': SystemCharacteristics(
-        '8x12x20',
-        480,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3840',
-    ),
-    'v5p-3968': SystemCharacteristics(
-        '4x4x124',
-        496,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-3968',
-    ),
-    'v5p-4096': SystemCharacteristics(
-        '8x16x16',
-        512,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4096',
-    ),
-    'v5p-4224': SystemCharacteristics(
-        '4x12x44',
-        528,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4224',
-    ),
-    'v5p-4352': SystemCharacteristics(
-        '4x8x68',
-        544,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4352',
-    ),
-    'v5p-4480': SystemCharacteristics(
-        '4x20x28',
-        560,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4480',
-    ),
-    'v5p-4608': SystemCharacteristics(
-        '12x12x16',
-        576,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4608',
-    ),
-    'v5p-4736': SystemCharacteristics(
-        '4x4x148',
-        592,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4736',
-    ),
-    'v5p-4864': SystemCharacteristics(
-        '4x8x76',
-        608,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4864',
-    ),
-    'v5p-4992': SystemCharacteristics(
-        '4x12x52',
-        624,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-4992',
-    ),
-    'v5p-5120': SystemCharacteristics(
-        '8x16x20',
-        640,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-5120',
-    ),
-    'v5p-5248': SystemCharacteristics(
-        '4x4x164',
-        656,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-5248',
-    ),
-    'v5p-5376': SystemCharacteristics(
-        '8x12x28',
-        672,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-5376',
-    ),
-    'v5p-5504': SystemCharacteristics(
-        '4x4x172',
-        688,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-5504',
-    ),
-    'v5p-5632': SystemCharacteristics(
-        '8x8x44',
-        704,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-5632',
-    ),
-    'v5p-5760': SystemCharacteristics(
-        '12x12x20',
-        720,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-5760',
-    ),
-    'v5p-5888': SystemCharacteristics(
-        '4x8x92',
-        736,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-5888',
-    ),
-    'v5p-6016': SystemCharacteristics(
-        '4x4x188',
-        752,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6016',
-    ),
-    'v5p-6144': SystemCharacteristics(
-        '12x16x16',
-        768,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6144',
-    ),
-    'v5p-6272': SystemCharacteristics(
-        '4x28x28',
-        784,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6272',
-    ),
-    'v5p-6400': SystemCharacteristics(
-        '8x20x20',
-        800,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6400',
-    ),
-    'v5p-6528': SystemCharacteristics(
-        '4x12x68',
-        816,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6528',
-    ),
-    'v5p-6656': SystemCharacteristics(
-        '8x8x52',
-        832,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6656',
-    ),
-    'v5p-6784': SystemCharacteristics(
-        '4x4x212',
-        848,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6784',
-    ),
-    'v5p-6912': SystemCharacteristics(
-        '12x12x24',
-        864,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-6912',
-    ),
-    'v5p-7040': SystemCharacteristics(
-        '4x20x44',
-        880,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7040',
-    ),
-    'v5p-7168': SystemCharacteristics(
-        '8x16x28',
-        896,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7168',
-    ),
-    'v5p-7296': SystemCharacteristics(
-        '4x12x76',
-        912,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7296',
-    ),
-    'v5p-7424': SystemCharacteristics(
-        '4x8x116',
-        928,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7424',
-    ),
-    'v5p-7552': SystemCharacteristics(
-        '4x4x236',
-        944,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7552',
-    ),
-    'v5p-7680': SystemCharacteristics(
-        '12x16x20',
-        960,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7680',
-    ),
-    'v5p-7808': SystemCharacteristics(
-        '4x4x244',
-        976,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7808',
-    ),
-    'v5p-7936': SystemCharacteristics(
-        '4x8x124',
-        992,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-7936',
-    ),
-    'v5p-8064': SystemCharacteristics(
-        '12x12x28',
-        1008,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8064',
-    ),
-    'v5p-8192': SystemCharacteristics(
-        '16x16x16',
-        1024,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8192',
-    ),
-    'v5p-8320': SystemCharacteristics(
-        '4x20x52',
-        1040,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8320',
-    ),
-    'v5p-8448': SystemCharacteristics(
-        '8x12x44',
-        1056,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8448',
-    ),
-    'v5p-8704': SystemCharacteristics(
-        '8x8x68',
-        1088,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8704',
-    ),
-    'v5p-8832': SystemCharacteristics(
-        '4x12x92',
-        1104,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8832',
-    ),
-    'v5p-8960': SystemCharacteristics(
-        '8x20x28',
-        1120,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-8960',
-    ),
-    'v5p-9216': SystemCharacteristics(
-        '12x16x24',
-        1152,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-9216',
-    ),
-    'v5p-9472': SystemCharacteristics(
-        '4x8x148',
-        1184,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-9472',
-    ),
-    'v5p-9600': SystemCharacteristics(
-        '12x20x20',
-        1200,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-9600',
-    ),
-    'v5p-9728': SystemCharacteristics(
-        '8x8x76',
-        1216,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-9728',
-    ),
-    'v5p-9856': SystemCharacteristics(
-        '4x28x44',
-        1232,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-9856',
-    ),
-    'v5p-9984': SystemCharacteristics(
-        '8x12x52',
-        1248,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-9984',
-    ),
-    'v5p-10240': SystemCharacteristics(
-        '16x16x20',
-        1280,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-10240',
-    ),
-    'v5p-10368': SystemCharacteristics(
-        '12x12x36',
-        1296,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-10368',
-    ),
-    'v5p-10496': SystemCharacteristics(
-        '4x8x164',
-        1312,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-10496',
-    ),
-    'v5p-10752': SystemCharacteristics(
-        '12x16x28',
-        1344,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-10752',
-    ),
-    'v5p-10880': SystemCharacteristics(
-        '4x20x68',
-        1360,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-10880',
-    ),
-    'v5p-11008': SystemCharacteristics(
-        '4x8x172',
-        1376,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-11008',
-    ),
-    'v5p-11136': SystemCharacteristics(
-        '4x12x116',
-        1392,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-11136',
-    ),
-    'v5p-11264': SystemCharacteristics(
-        '8x16x44',
-        1408,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-11264',
-    ),
-    'v5p-11520': SystemCharacteristics(
-        '12x20x24',
-        1440,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-11520',
-    ),
-    'v5p-11648': SystemCharacteristics(
-        '4x28x52',
-        1456,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-11648',
-    ),
-    'v5p-11776': SystemCharacteristics(
-        '8x8x92',
-        1472,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-11776',
-    ),
-    'v5p-11904': SystemCharacteristics(
-        '4x12x124',
-        1488,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-11904',
-    ),
-    'v5p-12032': SystemCharacteristics(
-        '4x8x188',
-        1504,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-12032',
-    ),
-    'v5p-12160': SystemCharacteristics(
-        '4x20x76',
-        1520,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-12160',
-    ),
-    'v5p-12288': SystemCharacteristics(
-        '16x16x24',
-        1536,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-12288',
-    ),
-    'v5p-13824': SystemCharacteristics(
-        '12x24x24',
-        1728,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-13824',
-    ),
-    'v5p-17920': SystemCharacteristics(
-        '16x20x28',
-        2240,
-        'tpu-v5p-slice',
-        'ct5p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5p-17920',
-    ),
-    # v5litepod
-    'v5litepod-16': SystemCharacteristics(
-        '4x4',
-        4,
-        'tpu-v5-lite-podslice',
-        'ct5lp-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5litepod-16',
-    ),
-    'v5litepod-32': SystemCharacteristics(
-        '4x8',
-        8,
-        'tpu-v5-lite-podslice',
-        'ct5lp-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5litepod-32',
-    ),
-    'v5litepod-64': SystemCharacteristics(
-        '8x8',
-        16,
-        'tpu-v5-lite-podslice',
-        'ct5lp-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5litepod-64',
-    ),
-    'v5litepod-128': SystemCharacteristics(
-        '8x16',
-        32,
-        'tpu-v5-lite-podslice',
-        'ct5lp-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5litepod-128',
-    ),
-    'v5litepod-256': SystemCharacteristics(
-        '16x16',
-        64,
-        'tpu-v5-lite-podslice',
-        'ct5lp-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v5litepod-256',
-    ),
-    # v4
-    'v4-8': SystemCharacteristics(
-        '2x2x1',
-        1,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-8',
-    ),
-    'v4-16': SystemCharacteristics(
-        '2x2x2',
-        2,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-16',
-    ),
-    'v4-32': SystemCharacteristics(
-        '2x2x4',
-        4,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-32',
-    ),
-    'v4-64': SystemCharacteristics(
-        '2x4x4',
-        8,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-64',
-    ),
-    'v4-128': SystemCharacteristics(
-        '4x4x4',
-        16,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-128',
-    ),
-    'v4-256': SystemCharacteristics(
-        '4x4x8',
-        32,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-256',
-    ),
-    'v4-512': SystemCharacteristics(
-        '4x8x8',
-        64,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-512',
-    ),
-    'v4-1024': SystemCharacteristics(
-        '8x8x8',
-        128,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-1024',
-    ),
-    'v4-1536': SystemCharacteristics(
-        '8x8x12',
-        192,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-1536',
-    ),
-    'v4-2048': SystemCharacteristics(
-        '8x8x16',
-        256,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-2048',
-    ),
-    'v4-4096': SystemCharacteristics(
-        '8x16x16',
-        512,
-        'tpu-v4-podslice',
-        'ct4p-hightpu-4t',
-        4,
-        AcceleratorType['TPU'],
-        'v4-4096',
-    ),
-    # CPU system characteristics
-    # m1-megamem-96-$VMs
-    'm1-megamem-96-1': SystemCharacteristics(
-        'N/A',
-        1,
-        'N/A',
-        'm1-megamem-96',
-        1,
-        AcceleratorType['CPU'],
-        'm1-megamem-96-1',
-    ),
-    # n2-standard-64-$VMs
-    'n2-standard-64-1': SystemCharacteristics(
-        'N/A',
-        1,
-        'N/A',
-        'n2-standard-64',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-64-1',
-    ),
-    # n2-standard-32-$VMs
-    'n2-standard-32-1': SystemCharacteristics(
-        'N/A',
-        1,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-1',
-    ),
-    'n2-standard-32-2': SystemCharacteristics(
-        'N/A',
-        2,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-2',
-    ),
-    'n2-standard-32-4': SystemCharacteristics(
-        'N/A',
-        4,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-4',
-    ),
-    'n2-standard-32-8': SystemCharacteristics(
-        'N/A',
-        8,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-8',
-    ),
-    'n2-standard-32-16': SystemCharacteristics(
-        'N/A',
-        16,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-16',
-    ),
-    'n2-standard-32-32': SystemCharacteristics(
-        'N/A',
-        32,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-32',
-    ),
-    'n2-standard-32-64': SystemCharacteristics(
-        'N/A',
-        64,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-64',
-    ),
-    'n2-standard-32-128': SystemCharacteristics(
-        'N/A',
-        128,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-128',
-    ),
-    'n2-standard-32-256': SystemCharacteristics(
-        'N/A',
-        256,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-256',
-    ),
-    'n2-standard-32-512': SystemCharacteristics(
-        'N/A',
-        512,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-512',
-    ),
-    'n2-standard-32-1024': SystemCharacteristics(
-        'N/A',
-        1024,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-1024',
-    ),
-    'n2-standard-32-2048': SystemCharacteristics(
-        'N/A',
-        2048,
-        'N/A',
-        'n2-standard-32',
-        1,
-        AcceleratorType['CPU'],
-        'n2-standard-32-2048',
-    ),
-}
-""" If you modify UserFacingNameToSystemCharacteristics you should also modify
-the corresponding Map in MaxText/accelerator_to_spec_map.py """
-# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 
 PathwaysExpectedInstancesMap = {
@@ -4273,27 +3024,6 @@ def get_gke_node_pool_version(
   return 0, node_pool_gke_version
 
 
-################### Subcommand Functions ###################
-def default_subcommand_function(
-    _args,
-) -> int:  # args is unused, so pylint: disable=invalid-name
-  """Default subcommand function.
-
-  Args:
-    _args: user provided arguments for running the command.
-
-  Returns:
-    0 if successful and 1 otherwise.
-  """
-  xpk_utils.xpk_print(
-      'Welcome to XPK! See below for overall commands:', flush=True
-  )
-  parser.print_help()
-  cluster_parser.print_help()
-  workload_parser.print_help()
-  return 0
-
-
 def cluster_create_pathways(args) -> None:
   """Function around cluster creation for Pathways.
 
@@ -5718,25 +4448,6 @@ def get_cpu_affinity(accelerator_type) -> str:
   return ''
 
 
-def get_system_characteristics(
-    args,
-) -> tuple[SystemCharacteristics | None, int]:
-  """Get system characteristics based on user provided arguments.
-
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    Tuple with string with the system characteristics and
-    int of 0 if successful and 1 otherwise.
-  """
-  device_type = args.tpu_type if args.tpu_type else args.device_type
-  if device_type in UserFacingNameToSystemCharacteristics:
-    return UserFacingNameToSystemCharacteristics[device_type], 0
-  else:
-    return None, 1
-
-
 def is_autoprovisioning_enabled(
     args, system: SystemCharacteristics
 ) -> tuple[bool, int]:
@@ -6909,1253 +5620,3 @@ def inspector(args) -> None:
         f' {final_return_code}'
     )
   xpk_utils.xpk_exit(final_return_code)
-
-
-def add_shared_arguments(custom_parser):
-  """Add shared arguments to the parser.
-
-  Args:
-    custom_parser: parser to add shared arguments to.
-  """
-  custom_parser.add_argument(
-      '--project',
-      type=str,
-      default=None,
-      help='GCE project name, defaults to "gcloud config project."',
-  )
-  custom_parser.add_argument(
-      '--zone',
-      type=str,
-      default=None,
-      help=(
-          'GCE zone, e.g. us-central2-b, defaults to "gcloud config '
-          'compute/zone." Only one of --zone or --region is allowed in a '
-          'command.'
-      ),
-  )
-  custom_parser.add_argument(
-      '--dry-run',
-      type=bool,
-      action=argparse.BooleanOptionalAction,
-      default=False,
-      help=(
-          'If given `--dry-run`, xpk will print the commands it wants to run'
-          ' but not run them. This is imperfect in cases where xpk might'
-          ' branch based on the output of commands'
-      ),
-  )
-
-
-def add_shared_cluster_create_required_arguments(args_parsers):
-  """Add shared required arguments in cluster create and Pathways cluster create.
-
-  Args:
-      List of cluster create required arguments parsers
-  """
-  for custom_parser in args_parsers:
-    custom_parser.add_argument(
-        '--cluster',
-        type=str,
-        default=None,
-        help=(
-            'The name of the cluster. Will be used as the prefix for internal'
-            ' objects in the cluster.'
-        ),
-        required=True,
-    )
-
-
-def add_shared_cluster_create_optional_arguments(args_parsers):
-  """Add shared optional arguments in cluster create and Pathways cluster create.
-
-  Args:
-      List of cluster create optional arguments parsers
-  """
-  for custom_parser in args_parsers:
-    add_shared_arguments(custom_parser)
-    custom_parser.add_argument(
-        '--host-maintenance-interval',
-        type=str,
-        choices=['AS_NEEDED', 'PERIODIC'],
-        default='AS_NEEDED',
-        help='The maintenance policy of the cluster and respective clusters.',
-    )
-    custom_parser.add_argument(
-        '--gke-version',
-        type=str,
-        help=(
-            'The GKE version of the cluster and respective clusters. The'
-            ' default is determined dynamically based on RAPID channel'
-            ' recommended version.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--num-slices',
-        type=int,
-        default=1,
-        help='The number of slices to run the job on, defaults to 1.',
-        required=False,
-    )
-    custom_parser.add_argument(
-        '--pathways-gce-machine-type',
-        type=str,
-        default='n1-standard-32',
-        help='The CPU type for Pathways CPU nodepools',
-    )
-    custom_parser.add_argument(
-        '--default-pool-cpu-machine-type',
-        type=str,
-        default='e2-standard-16',
-        help=(
-            'Set the machine type within the default cpu node pool. For'
-            ' regional clusters, all zones must support the machine type.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--cluster-cpu-machine-type',
-        type=str,
-        default='',
-        help=(
-            'Getting deprecated soon! Please use'
-            ' --default-pool-cpu-machine-typeinstead, to denote the machine'
-            ' type of the default cpu node pool. Set the machine type of other'
-            ' cpu nodepools using --device-type.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--default-pool-cpu-num-nodes',
-        type=int,
-        default=6,
-        help=(
-            'Set the number of nodes within the default cpu node pool. This is'
-            ' set to 6 by default. Autoscaling is enabled to scale this value'
-            ' over time.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--custom-cluster-arguments',
-        type=str,
-        default='',
-        help=(
-            'Users can add their own arguments to customize their cluster'
-            ' create command. Do note, these will not override already used'
-            ' cluster creation arguments. e.g.'
-            " --custom-cluster-arguments='--network=mtu9k --subnetwork=mtu9k'"
-        ),
-    )
-    custom_parser.add_argument(
-        '--custom-nodepool-arguments',
-        type=str,
-        default='',
-        help=(
-            'Users can add their own arguments to customize their node pool '
-            ' create command. Do note, these will not override already used'
-            ' node pool creation arguments. e.g.'
-            ' --custom-nodepool-arguments="--disk-size=300"'
-        ),
-    )
-    custom_parser.add_argument(
-        '--force',
-        action='store_true',
-        help=(
-            'Forces node pool creation and delete commands to run without'
-            ' additional approval.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--custom-tpu-nodepool-arguments',
-        type=str,
-        default='',
-        help=(
-            'DEPRECATING SOON! Please use --custom-nodepool-arguments to'
-            ' customize node pool create command. Do note, these will not'
-            ' override already used node pool creation arguments. Example usage'
-            ' --custom-tpu-nodepool-arguments="--enable-ip-alias"'
-        ),
-    )
-
-
-def add_shared_cluster_create_tensorboard_arguments(args_parsers):
-  """Add shared tensorboard arguments in cluster create and Pathways cluster create.
-  Note that this feature enables non-Pathways workloads to use tensorboard arguments
-  on a Pathways cluster.
-  Args:
-      List of cluster create tensorboard arguments parsers
-  """
-  for custom_parser in args_parsers:
-    custom_parser.add_argument(
-        '--create-vertex-tensorboard',
-        action='store_true',
-        help='Set this flag to create a Tensorboard instance in Vertex AI.',
-    )
-    custom_parser.add_argument(
-        '--tensorboard-region',
-        type=str,
-        default='us-central1',
-        help=(
-            'The region to create Vertex Tensorboard instance in. Visit'
-            ' https://cloud.google.com/vertex-ai/docs/general/locations#available-regions'
-            ' to view regions supported by Vertex AI. By default, Tensorboard'
-            ' instance will be created in us-central1.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--tensorboard-name',
-        type=str,
-        required=False,
-        help=(
-            'The name of Vertex Tensorboard instance to create. '
-            'If not specified, a Tensorboard instance with the name '
-            f'<cluster>-{DEFAULT_VERTEX_TENSORBOARD_NAME} will be created.'
-        ),
-    )
-
-
-def add_shared_cluster_create_capacity_arguments(args_parsers):
-  """Add shared capacity arguments in cluster create and Pathways cluster create.
-
-  Args:
-      List of cluster create capacity arguments parsers
-  """
-  for custom_parser in args_parsers:
-    custom_parser.add_argument(
-        '--on-demand',
-        action='store_true',
-        help=(
-            'Sets node pool creation to use on-demand resources. '
-            ' See `--reservation` or `--spot` for other capacity types.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--reservation',
-        type=str,
-        help=(
-            'The reservation to be used for acquiring resources in the'
-            ' cluster. This will attempt to find the provided reservation.'
-            ' See `--spot` or `--on-demand` for other capacity types.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--spot',
-        action='store_true',
-        help=(
-            'Sets node pool creation to use spot resources.'
-            ' See `--reservation` or `--on-demand` for other capacity types.'
-        ),
-    )
-
-
-def add_shared_workload_create_required_arguments(args_parsers):
-  """Add shared required arguments in workload create and Pathways workload create.
-
-  Args:
-      List of workload create required arguments parsers
-  """
-  for custom_parser in args_parsers:
-    custom_parser.add_argument(
-        '--workload',
-        type=xpk_utils.workload_name_type,
-        default=None,
-        help='The name of the workload to run.',
-        required=True,
-    )
-    custom_parser.add_argument(
-        '--cluster',
-        type=str,
-        default=None,
-        help='The name of the cluster to run the job on.',
-        required=True,
-    )
-
-
-def add_shared_workload_create_optional_arguments(args_parsers):
-  """Add shared optional arguments in workload create and Pathways workload create.
-
-  Args:
-      List of workload create optional arguments parsers
-  """
-  for custom_parser in args_parsers:
-    add_shared_arguments(custom_parser)
-    custom_parser.add_argument(
-        '--docker-name',
-        type=str,
-        default='jax-tpu',
-        help=(
-            'The name of the docker-image to use, default and typically'
-            ' `jax-tpu`.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--num-slices',
-        type=int,
-        default=1,
-        help='The number of slices to use, default=1.',
-    )
-    custom_parser.add_argument(
-        '--priority',
-        type=str,
-        default='medium',
-        choices=['very-low', 'low', 'medium', 'high', 'very-high'],
-        help=(
-            'A priority, one of `very-low`, `low`, `medium`, `high` or'
-            ' `very-high`. Defaults to `medium`.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--max-restarts',
-        type=str,
-        default='0',
-        help=(
-            'Maximum number of times the JobSet will be restarted upon failure.'
-            ' Defaults to 0.'
-        ),
-    )
-    custom_parser.add_argument(
-        '-tgps',
-        '--termination-grace-period-seconds',
-        type=str,
-        default='30',
-        help=(
-            'Maximum wait time for a workload Pod to wrap up after a disruption'
-            ' event or deletion request.Defaults to 30 seconds.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--enable-debug-logs',
-        action='store_true',
-        help=(
-            'Set this flag to get verbose logging to investigate the issue in'
-            ' the workload.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--restart-on-user-code-failure',
-        action='store_true',
-        help=(
-            'Adding this argument will return user failures back to the jobset'
-            ' manager allowing restarts on user code when --max-restarts is set'
-            ' greater than 0. By default, this is not enabled, and workloads'
-            ' will not restart from user code failures. This is enabled by'
-            ' default on Pathways workloads.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--headless',
-        action='store_true',
-        help=(
-            'Please provide this argument to create Pathways workloads in'
-            ' headless mode. This arg can only be used in `xpk workload'
-            ' create-pathways`(preferred) or `xpk workload create'
-            ' --use-pathways.` (--use-pathways will be deprecated soon).'
-        ),
-    )
-    custom_parser.add_argument(
-        '--proxy-server-image',
-        type=str,
-        default=(
-            'us-docker.pkg.dev/cloud-tpu-v2-images/pathways/proxy_server:latest'
-        ),
-        help=(
-            'Please provide the proxy server image for Pathways. This arg can'
-            ' only be used in `xpk workload create-pathways`(preferred) or `xpk'
-            ' workload create --use-pathways.` (--use-pathways will be'
-            ' deprecated soon).'
-        ),
-    )
-    custom_parser.add_argument(
-        '--server-image',
-        type=str,
-        default='us-docker.pkg.dev/cloud-tpu-v2-images/pathways/server:latest',
-        help=(
-            'Please provide the server image for Pathways. This arg can only be'
-            ' used in `xpk workload create-pathways`(preferred) or `xpk'
-            ' workload create --use-pathways.` (--use-pathways will be'
-            ' deprecated soon).'
-        ),
-    )
-    custom_parser.add_argument(
-        '--pathways-gcs-location',
-        type=str,
-        default='gs://cloud-pathways-staging/tmp',
-        help=(
-            'Please provide the GCS location to store Pathways artifacts. This'
-            ' arg can only be used in `xpk workload create-pathways`(preferred)'
-            ' or `xpk workload create --use-pathways.` (--use-pathways will be'
-            ' deprecated soon).'
-        ),
-    )
-
-
-def add_shared_workload_create_env_arguments(args_parsers):
-  """Add shared workload create environment arguments in workload create and Pathways workload create.
-
-  Args:
-      List of workload create environment arguments parsers
-  """
-  for custom_parser in args_parsers:
-    workload_env_arguments = custom_parser.add_mutually_exclusive_group()
-    workload_env_arguments.add_argument(
-        '--env-file',
-        type=str,
-        default=None,
-        help=(
-            'Environment file to be applied to the container.  This file should'
-            ' use the syntax <variable>=value (which sets the variable to the'
-            ' given value) or <variable> (which takes the value from the local'
-            ' environment), and # for comments.'
-        ),
-    )
-    workload_env_arguments.add_argument(
-        '--env',
-        action='append',
-        type=str,
-        help=(
-            'Environment variable to set in the container environment. '
-            'The format is <variable>=value'
-        ),
-    )
-
-
-def add_shared_workload_base_docker_image_arguments(args_parsers):
-  """Add shared base docker image arguments in workload create and Pathways workload create.
-
-  Args:
-      List of workload create base docker image arguments parsers
-  """
-  for custom_parser in args_parsers:
-    custom_parser.add_argument(
-        '--base-docker-image',
-        type=str,
-        default=default_docker_image,
-        help=(
-            f'The base docker-image to use, default {default_docker_image}. If'
-            ' using a custom docker image it is typically addressed as'
-            ' gcr.io/${PROJECT}/${NAME}:latest. This docker image will be'
-            ' used as a base image by default and the `--script-dir` by'
-            ' default will be added to the image.'
-        ),
-    )
-    custom_parser.add_argument(
-        '--script-dir',
-        type=xpk_utils.directory_path_type,
-        default=default_script_dir,
-        help=(
-            'The local location of the directory to copy to the docker image'
-            ' and run the main command from. Defaults to current working'
-            ' directory.'
-        ),
-    )
-
-
-def add_shared_workload_docker_image_arguments(args_parsers):
-  """Add shared docker image arguments in workload create and Pathways workload create.
-
-  Args:
-      List of workload create docker image arguments parsers
-  """
-  for custom_parser in args_parsers:
-    custom_parser.add_argument(
-        '--docker-image',
-        type=str,
-        help=(
-            'The version of the docker-image to use. By default, '
-            ' `--base-docker-image` is used. Set this argument if the user'
-            ' wants the docker image to be used directly by the xpk workload. a'
-            ' custom docker image it is typically addressed as'
-            ' gcr.io/${PROJECT}/${NAME}:latest. This docker image will be used'
-            ' directly by the xpk workload.'
-        ),
-    )
-
-
-def add_shared_workload_create_tensorboard_arguments(args_parsers):
-  """Add shared tensorboard arguments in workload create and Pathways workload create.
-
-  Args:
-      List of workload create optional arguments parsers
-  """
-  for custom_parser in args_parsers:
-    custom_parser.add_argument(
-        '--use-vertex-tensorboard',
-        action='store_true',
-        help='Set this flag to view workload data on Vertex Tensorboard.',
-    )
-    custom_parser.add_argument(
-        '--experiment-name',
-        type=str,
-        required=False,
-        help=(
-            'The name of Vertex Experiment to create. '
-            'If not specified, a Vertex Experiment with the name '
-            '<cluster>-<workload> will be created.'
-        ),
-    )
-
-
-############### Define flags ###############
-# Create top level parser for xpk command.
-parser = argparse.ArgumentParser(description='xpk command', prog='xpk')
-
-xpk_subcommands = parser.add_subparsers(
-    title='xpk subcommands', dest='xpk_subcommands', help='Top level commands'
-)
-parser.set_defaults(func=default_subcommand_function)
-
-#### "cluster" command parser. ####
-cluster_parser = xpk_subcommands.add_parser(
-    'cluster',
-    help='Commands around creating, deleting, and viewing clusters.',
-)
-cluster_parser.set_defaults(func=default_subcommand_function)
-cluster_subcommands = cluster_parser.add_subparsers(
-    title='cluster subcommands',
-    dest='xpk_cluster_subcommands',
-    help=(
-        'These are commands related to cluster management. Look at help for'
-        ' specific subcommands for more details.'
-    ),
-)
-
-### "cluster create" command parser ###
-cluster_create_parser = cluster_subcommands.add_parser(
-    'create', help='Create cloud clusters.'
-)
-cluster_create_required_arguments = cluster_create_parser.add_argument_group(
-    'Required Arguments',
-    'Arguments required for cluster create.',
-)
-cluster_create_optional_arguments = cluster_create_parser.add_argument_group(
-    'Optional Arguments', 'Arguments optional for cluster create.'
-)
-cluster_create_capacity_arguments = cluster_create_parser.add_argument_group(
-    'Capacity Arguments', 'Arguments related to capacity for cluster create.'
-)
-cluster_create_tensorboard_arguments = cluster_create_parser.add_argument_group(
-    'Optional Vertex AI Tensorboard Arguments',
-    'Arguments for creating Vertex AI Tensorboard in cluster create.',
-)
-
-### Required arguments specific to "cluster create"
-
-cluster_device_group = (
-    cluster_create_required_arguments.add_mutually_exclusive_group(
-        required=True
-    )
-)
-cluster_device_group.add_argument(
-    '--tpu-type',
-    type=str,
-    default=None,
-    help='The tpu type to use, v5litepod-16, etc.',
-)
-cluster_device_group.add_argument(
-    '--device-type',
-    type=str,
-    default=None,
-    help=(
-        'The device type to use (can be tpu or gpu or cpu), v5litepod-16,'
-        ' h100-80gb-8, n2-standard-32-4 etc.'
-    ),
-)
-
-### Optional arguments specific to "cluster create"
-cluster_create_optional_arguments.add_argument(
-    '--num-nodes',
-    type=int,
-    default=2,
-    help='The number of nodes for a cluster, defaults to 2.',
-    required=False,
-)
-cluster_create_optional_arguments.add_argument(
-    '--enable-pathways',
-    action='store_true',
-    help=(
-        'DEPRECATING SOON!!! Please use `xpk cluster create-pathways`.'
-        ' Enable cluster to accept Pathways workloads.'
-    ),
-)
-
-### Autoprovisioning arguments specific to "cluster create"
-cluster_create_autoprovisioning_arguments = (
-    cluster_create_parser.add_argument_group(
-        'Optional Autoprovisioning Arguments',
-        'Arguments optional for enabling autoprovisioning.',
-    )
-)
-cluster_create_autoprovisioning_arguments.add_argument(
-    '--enable-autoprovisioning',
-    action='store_true',
-    help='Enable GKE features for autoprovisioning node pools in GKE clusters.',
-)
-cluster_create_autoprovisioning_arguments.add_argument(
-    '--autoprovisioning-min-chips',
-    type=int,
-    help=(
-        'Optionally set the minimum autoprovisioning accelerator resources in'
-        ' units of chips.By default, autoprovisioning will use the number of'
-        ' resources in the cluster as the minimum, and maximum.'
-    ),
-)
-cluster_create_autoprovisioning_arguments.add_argument(
-    '--autoprovisioning-max-chips',
-    type=int,
-    help=(
-        'Optionally set the maximum autoprovisioning accelerator resources in'
-        ' units of chips.By default, autoprovisioning will use the number of'
-        ' resources in the cluster as the minimum, and maximum.'
-    ),
-)
-
-
-### "cluster create-pathways" command parser ###
-
-cluster_create_pathways_parser = cluster_subcommands.add_parser(
-    'create-pathways',
-    help='Create Pathways-on-Cloud clusters.',
-)
-cluster_create_pathways_required_arguments = (
-    cluster_create_pathways_parser.add_argument_group(
-        'Required Arguments',
-        'Arguments required for cluster create-pathways.',
-    )
-)
-cluster_create_pathways_optional_arguments = (
-    cluster_create_pathways_parser.add_argument_group(
-        'Optional Arguments', 'Arguments optional for cluster create-pathways.'
-    )
-)
-cluster_create_pathways_capacity_arguments = (
-    cluster_create_pathways_parser.add_argument_group(
-        'Capacity Arguments',
-        'Arguments related to capacity for cluster create-pathways.',
-    )
-)
-cluster_create_pathways_tensorboard_arguments = (
-    cluster_create_pathways_parser.add_argument_group(
-        'Optional Vertex AI Tensorboard Arguments',
-        'Arguments for creating Vertex AI Tensorboard in cluster create.',
-    )
-)
-
-### Pathways required arguments specific to "cluster create"
-cluster_create_pathways_required_arguments.add_argument(
-    '--tpu-type',
-    type=str,
-    default=None,
-    help='The tpu type to use, v5litepod-16, etc.',
-)
-
-
-add_shared_cluster_create_required_arguments([
-    cluster_create_required_arguments,
-    cluster_create_pathways_required_arguments,
-])
-add_shared_cluster_create_optional_arguments([
-    cluster_create_optional_arguments,
-    cluster_create_pathways_optional_arguments,
-])
-add_shared_cluster_create_capacity_arguments([
-    cluster_create_capacity_arguments,
-    cluster_create_pathways_capacity_arguments,
-])
-add_shared_cluster_create_tensorboard_arguments([
-    cluster_create_tensorboard_arguments,
-    cluster_create_pathways_tensorboard_arguments,
-])
-
-cluster_create_parser.set_defaults(func=cluster_create)
-cluster_create_pathways_parser.set_defaults(func=cluster_create_pathways)
-
-
-### "cluster delete" command parser ###
-cluster_delete_parser = cluster_subcommands.add_parser(
-    'delete',
-    help='Delete cloud clusters.',
-)
-cluster_delete_required_arguments = cluster_delete_parser.add_argument_group(
-    'Required Arguments',
-    'Arguments required for cluster delete.',
-)
-cluster_delete_optional_arguments = cluster_delete_parser.add_argument_group(
-    'Optional Arguments', 'Arguments optional for cluster delete.'
-)
-
-### Required arguments
-cluster_delete_required_arguments.add_argument(
-    '--cluster',
-    type=str,
-    default=None,
-    help='The name of the cluster to be deleted.',
-    required=True,
-)
-
-### Optional Arguments
-add_shared_arguments(cluster_delete_optional_arguments)
-cluster_delete_parser.set_defaults(func=cluster_delete)
-
-### "cluster cacheimage" command parser ###
-cluster_cacheimage_parser = cluster_subcommands.add_parser(
-    'cacheimage',
-    help='Cache image.',
-)
-cluster_cacheimage_required_arguments = (
-    cluster_cacheimage_parser.add_argument_group(
-        'Required Arguments',
-        'Arguments required for cluster cacheimage.',
-    )
-)
-cluster_cacheimage_optional_arguments = (
-    cluster_cacheimage_parser.add_argument_group(
-        'Optional Arguments', 'Arguments optional for cluster cacheimage.'
-    )
-)
-cluster_cacheimage_group = (
-    cluster_cacheimage_parser.add_mutually_exclusive_group(required=True)
-)
-
-### Device Type Argument
-cluster_cacheimage_group.add_argument(
-    '--tpu-type',
-    type=str,
-    default=None,
-    help='The tpu type to cache images on, v5litepod-16, etc.',
-)
-cluster_cacheimage_group.add_argument(
-    '--device-type',
-    type=str,
-    default=None,
-    help=(
-        'The device type to cache images on (can be tpu or gpu), v5litepod-16,'
-        ' h100-80gb-8, etc.'
-    ),
-)
-
-### Required arguments
-cluster_cacheimage_required_arguments.add_argument(
-    '--cluster',
-    type=str,
-    default=None,
-    help='The name of the cluster to cache the image.',
-    required=True,
-)
-cluster_cacheimage_required_arguments.add_argument(
-    '--docker-image',
-    type=str,
-    default=None,
-    help='The docker-image to cache.',
-    required=True,
-)
-
-### Optional Arguments
-add_shared_arguments(cluster_cacheimage_optional_arguments)
-cluster_cacheimage_optional_arguments.add_argument(
-    '--cache-key',
-    type=str,
-    default='containerimage',
-    help='The key to cache the docker image under.',
-    required=False,
-)
-cluster_cacheimage_parser.set_defaults(func=cluster_cacheimage)
-
-### "cluster describe" command parser ###
-cluster_describe_parser = cluster_subcommands.add_parser(
-    'describe',
-    help='Describe a cluster.',
-)
-cluster_describe_required_arguments = (
-    cluster_describe_parser.add_argument_group(
-        'Required Arguments',
-        'Arguments required for cluster describe.',
-    )
-)
-cluster_describe_optional_arguments = (
-    cluster_describe_parser.add_argument_group(
-        'Optional Arguments', 'Arguments optional for cluster describe.'
-    )
-)
-
-### Required arguments
-cluster_describe_required_arguments.add_argument(
-    '--cluster',
-    type=str,
-    default=None,
-    help='The name of the cluster to be describe.',
-    required=True,
-)
-### Optional Arguments
-add_shared_arguments(cluster_describe_optional_arguments)
-
-
-cluster_describe_parser.set_defaults(func=cluster_describe)
-
-# "cluster list" command parser.
-cluster_list_parser = cluster_subcommands.add_parser(
-    'list', help='List cloud clusters.'
-)
-cluster_list_optional_arguments = cluster_list_parser.add_argument_group(
-    'Optional Arguments', 'Arguments optional for cluster list.'
-)
-### Optional Arguments
-add_shared_arguments(cluster_list_optional_arguments)
-
-
-cluster_list_parser.set_defaults(func=cluster_list)
-
-#### "workload" command parser. ####
-workload_parser = xpk_subcommands.add_parser(
-    'workload', help='commands around workload management'
-)
-
-workload_parser.set_defaults(func=default_subcommand_function)
-workload_subcommands = workload_parser.add_subparsers(
-    title='workload subcommands',
-    dest='xpk_workload_subcommands',
-    help=(
-        '`create`, `create-pathways`, `list` and `delete` workloads on clusters'
-    ),
-)
-
-# "workload create" command parser.
-workload_create_parser = workload_subcommands.add_parser(
-    'create', help='Create a new job.'
-)
-workload_create_parser_required_arguments = (
-    workload_create_parser.add_argument_group(
-        'Workload Built-in Arguments',
-        'Configure xpk to create a Workload for you.',
-    )
-)
-workload_create_parser_optional_arguments = (
-    workload_create_parser.add_argument_group(
-        'Optional Arguments', 'Arguments optional for `workload create`.'
-    )
-)
-workload_base_docker_image_arguments = workload_create_parser.add_argument_group(
-    'Base Docker Image Arguments',
-    'User supplies a base image or by default the image is set by xpk.'
-    ' Xpk will add the `script_dir` to the base image creating an anonymous'
-    ' docker image. These arguments are exclusive to `--docker-image`.',
-)
-workload_docker_image_arguments = workload_create_parser.add_argument_group(
-    'Docker Image Arguments',
-    '`--base-docker-image` is used by default. Set this argument if the'
-    ' user wants the docker image to be used directly by the xpk workload.',
-)
-workload_create_autoprovisioning_arguments = (
-    workload_create_parser.add_argument_group(
-        'Optional Autoprovisioning Arguments',
-        'Arguments for configuring autoprovisioning.',
-    )
-)
-workload_pathways_workload_arguments = workload_create_parser.add_argument_group(
-    'Pathways Image Arguments',
-    'If --use-pathways is provided, user wants to set up a'
-    'Pathways workload on xpk.',
-)
-workload_vertex_tensorboard_arguments = (
-    workload_create_parser.add_argument_group(
-        'Vertex Tensorboard Arguments',
-        'Arguments for creating Vertex AI Experiment in workload create.',
-    )
-)
-
-### "workload create" Required arguments
-workload_create_parser_required_arguments.add_argument(
-    '--command',
-    type=str,
-    default=None,
-    help=(
-        'Main command to run on each VM. This script runs within the docker '
-        'container. Typically this looks like "--command=\'python3 train.py\'" '
-        'but if your docker container is missing the dependencies, it might '
-        'look more like "--command=\'bash setup.sh && python3 train.py\'".'
-    ),
-    required=True,
-)
-workload_device_group = (
-    workload_create_parser_required_arguments.add_mutually_exclusive_group(
-        required=True
-    )
-)
-workload_device_group.add_argument(
-    '--tpu-type',
-    type=str,
-    default=None,
-    help='The tpu type to use, v5litepod-16, etc.',
-)
-workload_device_group.add_argument(
-    '--device-type',
-    type=str,
-    default=None,
-    help=(
-        'The device type to use (can be tpu or gpu or cpu), v5litepod-16,'
-        ' h100-80gb-8, n2-standard-32-4 etc.'
-    ),
-)
-
-workload_create_parser_optional_arguments.add_argument(
-    '--num-nodes',
-    type=int,
-    default=1,
-    help='The number of nodes to use, default=1.',
-)
-workload_create_parser_optional_arguments.add_argument(
-    '--scheduler',
-    type=str,
-    default='default-scheduler',
-    help=(
-        'Which scheduler you want to use. Defaults to `default-scheduler`. If'
-        ' your cluster is configured for high throughput scheduling, you might'
-        ' want to use `gke.io/high-throughput-scheduler`.If your cluster is'
-        ' configured for topology-aware scheduling, you might want to use'
-        ' `gke.io/topology-aware-auto`.'
-    ),
-)
-workload_create_parser_optional_arguments.add_argument(
-    '--debug-dump-gcs',
-    type=str,
-    default=None,
-    help=(
-        'GCS bucket or a directory within a bucket, e.g gs://bucket/subdir, '
-        'where debugging information such as HLO dumps are uploaded'
-    ),
-)
-workload_create_parser_optional_arguments.add_argument(
-    '--deploy-stacktrace-sidecar',
-    action='store_true',
-    help=(
-        'Add this argument to deploy a sidecar container that will '
-        'read the stack traces collected in /tmp/debugging directory '
-        'and forward them to Cloud Logging for TPU workloads.'
-    ),
-)
-
-# Autoprovisioning workload arguments
-workload_create_autoprovisioning_arguments.add_argument(
-    '--on-demand',
-    action='store_true',
-    help=(
-        'Sets autoprovisioning to use on-demand resources for the workload'
-        ' request. See `--reservation` or `--spot` for other capacity types.'
-    ),
-)
-workload_create_autoprovisioning_arguments.add_argument(
-    '--reservation',
-    type=str,
-    help=(
-        'Sets autoprovisioning to use reservation resources for the workload'
-        ' request. This will attempt to find the provided reservation. See'
-        ' `--spot` or `--on-demand` for other capacity types.'
-    ),
-)
-workload_create_autoprovisioning_arguments.add_argument(
-    '--spot',
-    action='store_true',
-    help=(
-        'Sets autoprovisioning to use spot resources.'
-        ' See `--reservation` or `--on-demand` for other capacity types.'
-    ),
-)
-
-# Pathways workload arguments
-workload_pathways_workload_arguments.add_argument(
-    '--use-pathways',
-    action='store_true',
-    help=(
-        'DECRATING SOON!!! Please use `xpk workload create-pathways` instead.'
-        ' Provide this argument to create Pathways workloads.'
-    ),
-)
-
-
-# "workload create-pathways" command parser.
-workload_create_pathways_parser = workload_subcommands.add_parser(
-    'create-pathways', help='Create a new job.'
-)
-workload_create_pathways_parser_required_arguments = (
-    workload_create_pathways_parser.add_argument_group(
-        'Workload create-pathways Built-in Arguments',
-        'Configure xpk to create a Pathways Workload for you.',
-    )
-)
-workload_create_pathways_parser_optional_arguments = (
-    workload_create_pathways_parser.add_argument_group(
-        'Optional Arguments',
-        'Arguments optional for `workload create-pathways`.',
-    )
-)
-workload_create_pathways_base_docker_image_arguments = workload_create_pathways_parser.add_argument_group(
-    'Base Docker Image Arguments',
-    'User supplies a base image or by default the image is set by xpk.'
-    ' Xpk will add the `script_dir` to the base image creating an anonymous'
-    ' docker image. These arguments are exclusive to `--docker-image`.',
-)
-workload_create_pathways_docker_image_arguments = workload_create_pathways_parser.add_argument_group(
-    'Docker Image Arguments',
-    '`--base-docker-image` is used by default. Set this argument if the'
-    ' user wants the docker image to be used directly by the xpk workload.',
-)
-workload_create_pathways_vertex_tensorboard_arguments = (
-    workload_create_pathways_parser.add_argument_group(
-        'Vertex Tensorboard Arguments',
-        'Arguments for creating Vertex AI Experiment in workload create.',
-    )
-)
-
-### "workload create-pathways" Required arguments, specific to Pathways
-workload_create_pathways_parser_required_arguments.add_argument(
-    '--tpu-type',
-    type=str,
-    default=None,
-    help='The tpu type to use, v5litepod-16, etc.',
-)
-
-workload_create_pathways_parser_optional_arguments.add_argument(
-    '--command',
-    type=str,
-    default=None,
-    help=(
-        'Main command to run on each VM. This script runs within the docker '
-        'container. Typically this looks like "--command=\'python3 train.py\'" '
-        'but if your docker container is missing the dependencies, it might '
-        'look more like "--command=\'bash setup.sh && python3 train.py\'".'
-    ),
-    required=False,
-)
-
-add_shared_workload_create_required_arguments([
-    workload_create_parser_required_arguments,
-    workload_create_pathways_parser_required_arguments,
-])
-add_shared_workload_create_optional_arguments([
-    workload_create_parser_optional_arguments,
-    workload_create_pathways_parser_optional_arguments,
-])
-add_shared_workload_create_env_arguments([
-    workload_create_parser_optional_arguments,
-    workload_create_pathways_parser_optional_arguments,
-])
-add_shared_workload_base_docker_image_arguments([
-    workload_base_docker_image_arguments,
-    workload_create_pathways_base_docker_image_arguments,
-])
-add_shared_workload_docker_image_arguments([
-    workload_docker_image_arguments,
-    workload_create_pathways_docker_image_arguments,
-])
-add_shared_workload_create_tensorboard_arguments([
-    workload_vertex_tensorboard_arguments,
-    workload_create_pathways_vertex_tensorboard_arguments,
-])
-
-# Set defaults for both workload create and workload create-pathways after adding all shared args.
-workload_create_parser.set_defaults(func=workload_create)
-workload_create_pathways_parser.set_defaults(func=workload_create_pathways)
-
-# "workload delete" command parser.
-workload_delete_parser = workload_subcommands.add_parser(
-    'delete', help='Delete job.'
-)
-workload_delete_parser_required_arguments = (
-    workload_delete_parser.add_argument_group(
-        'Required Arguments',
-        'Arguments required for `job delete`.',
-    )
-)
-workload_delete_parser_optional_arguments = (
-    workload_delete_parser.add_argument_group(
-        'Optional Arguments', 'Arguments optional for `job delete`.'
-    )
-)
-add_shared_arguments(workload_delete_parser_optional_arguments)
-
-### "workload delete" Required arguments
-workload_delete_parser_required_arguments.add_argument(
-    '--cluster',
-    type=str,
-    default=None,
-    help='The name of the cluster to delete the job on.',
-    required=True,
-)
-### "workload delete" Optional arguments
-workload_delete_parser_optional_arguments.add_argument(
-    '--workload',
-    type=xpk_utils.workload_name_type,
-    default=None,
-    help=(
-        'The name of the workload to delete. If the workload is not specified, '
-        'all workloads will be deleted from the cluster.'
-    ),
-)
-workload_delete_parser_optional_arguments.add_argument(
-    '--filter-by-job',
-    type=str,
-    help=(
-        'Filters the arguments based on job name. Provide a regex expressionto'
-        ' parse jobs that match the pattern or provide a job name to delete a'
-        ' single job.'
-    ),
-)
-workload_delete_parser_optional_arguments.add_argument(
-    '--filter-by-status',
-    type=str,
-    default='EVERYTHING',
-    choices=[
-        'EVERYTHING',
-        'FINISHED',
-        'RUNNING',
-        'QUEUED',
-        'FAILED',
-        'SUCCESSFUL',
-    ],
-    help=(
-        'Filters the arguments based on status. Selected filters are listed'
-        ' above. FAILED and SUCCESSFUL are sub-states of FINISHED.'
-    ),
-    required=False,
-)
-workload_delete_parser_optional_arguments.add_argument(
-    '--force',
-    action='store_true',
-    help='Forces workload deletion command to run without additional approval.',
-)
-
-workload_delete_parser.set_defaults(func=workload_delete)
-
-# "workload list" command parser.
-workload_list_parser = workload_subcommands.add_parser(
-    'list', help='List jobs.'
-)
-
-workload_list_parser.add_argument(
-    '--cluster',
-    type=str,
-    default=None,
-    help='The name of the cluster to list jobs on.',
-    required=True,
-)
-
-workload_list_parser.add_argument(
-    '--filter-by-status',
-    type=str,
-    default='EVERYTHING',
-    choices=[
-        'EVERYTHING',
-        'FINISHED',
-        'RUNNING',
-        'QUEUED',
-        'FAILED',
-        'SUCCESSFUL',
-    ],
-    help=(
-        'Filters the arguments based on status. Selected filters are listed'
-        ' above. FAILED and SUCCESSFUL are sub-states of FINISHED.'
-    ),
-    required=False,
-)
-
-workload_list_parser.add_argument(
-    '--filter-by-job',
-    type=str,
-    help=(
-        'Filters the arguments based on job name. Provide a regex expressionto'
-        ' parse jobs that match the pattern or provide a job name to view a'
-        ' single job.'
-    ),
-    required=False,
-)
-
-workload_list_wait_for_job_completion_arguments = (
-    workload_list_parser.add_argument_group(
-        'Wait for Job Completion Arguments',
-        'Arguments for waiting on the completion of a job.',
-    )
-)
-
-workload_list_wait_for_job_completion_arguments.add_argument(
-    '--wait-for-job-completion',
-    type=str,
-    default=None,
-    help='The name of the job to wait on.',
-    required=False,
-)
-
-workload_list_wait_for_job_completion_arguments.add_argument(
-    '--timeout',
-    type=int,
-    default=None,
-    help=(
-        'Amount of time to wait for job in seconds. Default is the max wait'
-        ' time, 1 week.'
-    ),
-    required=False,
-)
-
-add_shared_arguments(workload_list_parser)
-
-workload_list_parser.set_defaults(func=workload_list)
-
-
-#### "inspector" command parser. ####
-inspector_parser = xpk_subcommands.add_parser(
-    'inspector',
-    help='commands around investigating workload, and Kueue failures.',
-)
-
-inspector_parser.set_defaults(func=default_subcommand_function)
-inspector_subcommands = inspector_parser.add_subparsers(
-    title='inspector subcommands',
-    dest='xpk_inspector_subcommands',
-    help='Investigate workload, and Kueue failures.',
-)
-
-inspector_parser_required_arguments = inspector_parser.add_argument_group(
-    'inspector Built-in Arguments', 'Arguments required for `inspector`.'
-)
-inspector_parser_optional_arguments = inspector_parser.add_argument_group(
-    'Optional Arguments', 'Arguments optional for `inspector`.'
-)
-
-### "inspector" Required arguments
-
-inspector_parser_required_arguments.add_argument(
-    '--cluster',
-    type=str,
-    default=None,
-    help='The name of the cluster to investigate.',
-    required=True,
-)
-
-### "inspector" Optional Arguments
-add_shared_arguments(inspector_parser_optional_arguments)
-
-inspector_parser_optional_arguments.add_argument(
-    '--workload',
-    type=xpk_utils.workload_name_type,
-    default=None,
-    help='The name of the workload to investigate.',
-)
-
-inspector_parser_optional_arguments.add_argument(
-    '--print-to-terminal',
-    action='store_true',
-    help=(
-        'Prints inspector output to terminal. A user can always look at the'
-        ' returned file.'
-    ),
-)
-
-inspector_parser.set_defaults(func=inspector)
-
-xpk_utils.xpk_print('Starting xpk', flush=True)
-main_args = parser.parse_args()
-main_args.func(main_args)
-
-
-################### Main ###################
-def main() -> None:
-  xpk_utils.xpk_print('XPK Done.', flush=True)
-
-
-if __name__ == '__main__':
-  main()
